pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpha-research-platform"
version = "0.1.0"
description = "Production-grade daily cross-sectional equities alpha research platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=2.2,<3",
  "pandas>=2.2,<3",
  "pydantic>=2.12,<3",
  "PyYAML>=6.0,<7",
  "pyarrow>=20,<21",
  "exchange-calendars>=4.13,<5",
]

[project.optional-dependencies]
research = [
  "scikit-learn>=1.7,<2",
  "jinja2>=3.1,<4",
  "matplotlib>=3.10,<4",
  "mlflow>=2.22,<3",
  "pandera>=0.26,<1",
]
dev = [
  "pytest>=9,<10",
  "ruff>=0.12,<1",
  "mypy>=1.18,<2",
  "pre-commit>=4.2,<5",
  "pip-tools>=7.4,<8",
]

[project.scripts]
alpha-research = "alpha_research.cli.main:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
